requirements.txt
langchain
langchain-ollama
langchain-chroma
pandas
pypdf
pdfplumber
PyMuPDF
pytesseract
Pillow
ibm-watsonx-ai
langchain-ibm
streamlit